Eriytä validaattorit omiin tiedostoihinsa. Tee uusi, parempi validateTiedostoInput.
