+++ backend/src/projekti/validator/validateAloituskuulutus.ts
@@ -1,0 +1,7 @@
+import { AloitusKuulutusInput } from "hassu-common/graphql/apiModel";
+import { DBProjekti } from "../../database/model";
+
+export function validateAloituskuulutus(_projekti: DBProjekti, _aloituskuulutus: AloitusKuulutusInput | null | undefined) {
+  //TODO
+  return;
+}
